learn: Add `//levy:blank` to question text
Add `//levy:blank` to question text. Lines with Evy source lines ending in
`//levy:blank` will not be presented in the rendered fill-in-the-blank answer
text. It is the students task to fill in these blanks. The full answer text
will be submitted and compared against the correct answer.

The text answer key, ie the correct answer, will still be generated using the
blanked lines.