learn: Expand on composition todo
Expand on composition todo: it seems that quite often questions should not be
used in the same quiz, even if not generated from the same txtar. Let's
facilitate this via an optional group field in the question Frontmatter. Only
one question per group may be selected per quiz or exercise. Add note to not
forget when reworking composition.