learn: Add sidebar to learn pages
Create sidebar, one per course with absolute URL paths, so that we can re-use
the same sidebar for any sub-entries of the course (unit, exercise,....).
Stop adding question as Individual HTML files, we will add them back in as
fragments in a future commit.
Create the sidebar structure similar to the sidebar of the doc-site and following the
course -> unit -> exercise hierarchy.