learn-site: Add HTML Fragments
Add HTML Fragments with extension to .htmlf to learn site so that we can use
them in our Vanilla-SPA (Vanilla JS, Single Page Application). In a follow-up
commit the Vanilla-SPA will only set the `<main>` elements innerHTML to the
contents of the fragment file, when clicking on link, rather than loading the
full `.html` page.

Remove `--skip-questions` flag from `levy` command, which was used to skip
individual answer exports as it is not needed anymore. We will have to always
export individual question .htmlf files.

Commit-group: 1/2