docs-site: Add HTML Fragments (generated)
Add HTML Fragments with extension to .htmlf to learn site so that we can use
them in our Vanilla-SPA (Vanilla JS, Single Page Application). In a follow-up
commit the Vanilla-SPA will only set the `<main>` elements innerHTML to the
contents of the fragment file, when clicking on link, rather than loading the
full `.html` page. We have already done this for the learn-site and are now
refactoring the docs-site as well to use the same mechanism.

Gen-command: make docs
Commit-group: 2/2