learn-site: Make content width a little narrower
Make content width a little narrower in response to watching my daughter Mali
struggle a little reading the fairly long lines with the default font size in
my default browser. Sneaking this in here even though it doesn't really
belong with the other commits, but didn't want to forget it.