deploy: Prepare docs site in separate public dir
Prepare docs site in separate public dir for deployment, so that we can
re-write relatives URL starting in `/`, which are needed for the single
sidebar. We are doing this exactly the same we we have already done it for
lean.evy.dev. This is a preparatory step for using fragments and spa.js with
the docs-site.