frontend: Fix mobile sidebar scroll bug (generated)
Fix mobile sidebar scroll bug where on mobile the header would also scroll
with the content of the sidebar. in order to fix this we had to wrap the
sidebar content (all but header) in a separate `<div>` which gets set to
scroll.

Gen-command: make docs && make learn && make snaps USE_DOCKER=1
Commit-group: 2/2