🧪 lab-site: Initialize lab site (#394)
Initialize lab site and lab.evy.dev subdomain. This is in preparation for a
hands-on lesson with a Melbourne primary school class soon to follow. The lab
subdomain currently hosts a straight up copy of play.evy.dev (symlinks on the
file system). The only difference is the the sample selection intended for
the first, introductory lesson.

Pull-request: https://github.com/evylang/evy/pull/394