play-site: Rework sample selection
Rework sample selection according to design specification by @jasonstrachan
(finally). This automatically flows on to lab challenges and should make
navigation through them a little easier.

Update a couple of the playground sample titles so that the fit in the sample
title box without ellipses for the default browser font size (16px).

Update end-to-end test snapshots.

Link: https://www.figma.com/design/4Wf4Mbq8I2Yxh2bptOnSE9/evy?node-id=1243-1340&t=vYTNo7m81uXbVvrG-0