🚤 play-site: Improve initialization concurrency (#401)
Improve initialization concurrency, so that we load sample.json, sample content
in parallel to the large evy.wasm file. The difference can be observed in
browser developer tools with network throttling enabled and watching the
waterfall - instead of the previous:

	|--- evy.wasm -------------------------|
	                                       |-- sample.json -|
	                                                        |-- welcome.evy -|

we now have

	|--- evy.wasm -------------------------|
	|-- sample.json -|
	                 |-- welcome.evy -|

This means that for any hash other than #welcome the editor contents are visible
and syntax colored faster.

Pull-request: https://github.com/evylang/evy/pull/401