lab-site: Change index.html symlink to a copy
Change index.html symlink to frontend/play/index.html to a copy of
frontend/play/index.html. In a follow-up commit we will change the
lab/index.html to contain some css overrides and to initialize to the lab
`#welcome` contents. We will also add a placeholder for the notes - so we need
a separate index.html to the one of the play site. However, the index.js and
index.css we keep as symlinks for the time being.