lab-site: Remove lines sample from overview
Remove lines sample from overview as we didn't get through it in class. When Tom
does the video recording it shouldn't show up either. We'll leave source and
notes in the repo for now, in case people used the URL and also to move it in
place we do a "more loops" lab (probably after conditionals).