lab: Create `if` statement lab
Create `if` statement lab. This is our third lab. I've tested it with my
daughter Mali (thank you!). She managed to do pretty much all of it by herself
with the code hints. The Monte Carlo Pi simulation was the trickiest I think,
but that's more intended as a walk through on the facilitators side anyway.

The content is on the wordy side but I intend to add some [[ NEXT ]] buttons
within the lab notes in the follow up commit.