build-tools: Rename md to docsite-gen
Rename md to docsite-gen as we are about to add a labsite-gen tool. This naming
also better aligns with site-gen that is is added as a final step. Maybe later
we should add them all into a single tool maybe even evy proper (full build),
but for now that should put things in a slightly better and more consistently
named place.