bytecode: Implement local variables
Add scopes to the compiler so that it can differentiate whether a
variable is global or local. A local variable is one that is not defined
at the top-level of the program - i.e. it is either in a block at the
top-level of the program or in a function (although functions are not
yet implemented in the compiler).

Space is allocated at the top of the stack for the locals needed, with
the stack pointer initialised to start after that. The space allocated
for locals is reused where it can be, for example, when two blocks at
the same level have a local variable, they can share the same space as
they both cannot be in scope at the same time.