builtins: Enforce sleep lower bound in JS runtime
When calling `sleep` in the JS runtime, enforce a minimum allowed
interval of 1 millisecond.

This change is motivated by watching students use Evy. I found they
experimented with very large and very small numbers. When writing a loop
with 10,000,000,000+ iterations containing `sleep 0`, their browser tabs
crashed. For example, run this program against `main`:

	i:num

	clear "blue"
	while i < 100000
	    move i 50
	    color "yellow"
	    circle 5
	    i = i + 0.1
	    sleep 0
	end

This is enough to freeze the browser tab on my PC for a few seconds, and
the Evy UI is unresponsive even after the render finishes. In my branch,
it runs smoothly - and slowly - with the STOP button still functional.

I experimented with a few values and found a minimum sleep of 1
millisecond (0.001 seconds) is enough to ensure that the browser remains
responsive.