builtin: Rename evy builtin `assert` to `test` (1/3)
Rename evy builtin `assert` to `test` as test seems to be a more
intuitive name than assert. Full disclosure @camh- said all along it
should `test`, but I felt the concept of a test is an abstraction level
higher, a collection of many asserts or so. I had to rename all
human-eval samples from `func test` to `func testAll`.

Still it seems with a bit of distance and reading the builtin docs
afresh: test is the more intuitive name.

Commit-group: 1/3