e2e: Reduce dialog snapshot test flakiness
Reduce dialog snapshot test flakiness, or at least attempt it. We just
add a 1 sec wait before we take the dialog snapshot which seems to have
removed any local flakiness. I'm really unsure what the right way to
fix this would be, but snapshot tests have been useful (and painful) in
the past and found real issues so we'll stick with them for now.