playsite: Fix share dialog primary button
The primary button in the share dialog said Done, without having done
the main action of sharing: copying the URL. We have changed it to say
`Copy`, copying the URL. If you don't wish to copy the URL, just hit
the close button of the dialog. As a result we had to tweak JS, CSS and
the snapshot tests.