labsite: Add tooling for Next button
Add markdown tooling for Next button generation from

	[Next]

A `<details>...</details>` HTMLBlock generated from Markdown also stops when a
`[Next]` button is found (in addition to thematic break and heading as end
marker).

As we were developing this we uncovered a bug around newlines inside code blocks
after conversion to HTMLBlock used again as MD. I've snuck in a fix and test
for that too.