labsite: Update JS and CSS for Next button
Add a bit new CSS and JS to the labsite (JS is shared with playsite via
symlink), that hides all Notes content after the first Next button on load and
only reveals content until the next button.

It also tweaks styling for the `<details>` blocks to align with the Next
buttons.

Sneak in a few other CSS updates. Never ending story.