labsite: Update contents to align with Next button (1/2)
We have removed verbose Headings for `<details>` section to fit on button.
This way details summary and next button are visually aligned.

We've also added to each section revealed by a Next button click a Heading.

As per usual I've reworded the content a little to make it more consistent with
how the labs have evolved. Hints are a little more elaborate now.

Commit-group: 1/2