labsite: Improve scrolling on Next click
Improve the scrolling experience on Next button click: We now always scroll the
next button to the top of the page with the top edge a little clipped. Smooth
scrolling is also enabled, so it should hopefully be more obvious what's going
on when clicking on the next button. Unfortunately smooth scrolling doesn't
seem to work on my linux machine in Firefox or Chrome, but it seem ok on MacOS
and iOS.