gallery: Add gallery to frontend and deployment
Add frontend/gallery with just a simple index.html redirect to
https://github.com/evylang/evy/wiki/Gallery. This is a preparatory step
for setting up a proper gallery. I want to get rid of links to the
github wiki all over the repo in a follow up step.

Add the gallery deployment to the firebase-deploy script and firebase
config files.

DNS setup to the new firebase hosting project has already been done on
the registrars for

- evy.dev
- evystage.dev
- evytest.dev