labsite: Fix reloading for labs with edits
When you edit a lab and reload the page you land on #welcome where the
editor is hidden. This means it appears on reload editor contents are
lost on the labsite.

As soon as we edit the editor contents we remove the URL fragment. This
has been done so you can easily go back to the "original" contents for
a sample on the playground or a challenge on the labsite.

However, if we reload an edited lab challenge we get "re-directed"
(inside JS) to the #welcome page as we have no URL fragment. #welcome
happens to have the editor hidden. The editor is _is_ in fact
repopulated on reload with the last content saved in SessionStorage,
just not displayed.

This logic worked for the playsite where we don't hide editors and have
no notes, however, on the labsite we probably want to keep
the "current" lab, repopulate the current lab-notes and the editor on reload.

To do this we store the "currentSample" in the sessionStorage alongside
the editor contents.

If set we reload the notes for it.

The code is becoming increasingly patchy - it seems it really is time to
factor out the editor and evy runtime logic and piece together playsite
and labsite separately. However, for the time being we'll work with
this.