github: Add check-uptodate to CI
Add `check-uptodate` to CI as a separate GitHub Job and remove it as
part of the `test` GitHub Job in order to speed things up.

The environment variable `TERM=vt100` needs to be set on CI for the
`doctest` make target which is a dependency of the `check-uptodate`
target. `doctest` runs evy code, including the `cls` command which
hangs if TERM is not set.