make: Rework make targets for parallelism
Rework make targets for parallelism, so that we can run

	make ci -j $(nproc)

to speed up the CI builds locally and on GitHub Actions. This required a
little of renaming of files used for markdown generation. It also seems
that the node tool prettier and stylelint cannot be installed / run via
npx in parallel.