make: Fix ci-check-uptodate for parallelism
Fix `check-uptodate` for parallelism. It seems Node is still stepping
on its toes when `check-prettier` and `check-style` are run in parallel
because `check-style` runs `npm ci` which delete `node_modules`.

While fixing this I ended up refactoring a bunch in the Makefile to
make the GitHub Workflow and the Makefile more aligned.

Add a `docker-pull` target to speed up the `deploy` job.