labsite: Use YouTube screenshot rather than embed
Use YouTube screenshot rather than embed, because YouTube embed despite
best efforts triggers warnings about third party cookies not going to
be supported in the future in Chrome. 250+ of them, even if explicitly
using the youtube-nocookie.com domain. So replace the embedded YouTube
player with a an image link that opens the video in new tab.