🐞 docsite: Fix broken PDF links (#433)
Fix broken PDF links which showed some really strange results on
initial click to Cheat Sheet or Evy Paper. This was caused by link
re-writing in the SPA module. All relative links are re-written on the
docs page to load the .htmlf fragment and display inside the `<main>`
area of the docs. This happens indiscriminately of file type, so the
bug was that we tried to load evy-cheat-sheet.pdff (not the double "f"
at the end). Patch it by using an additional filter condition that only
re-writes relative links that end in `.html`.

Admittedly, this is a bit of a foot-gun, I guess this can happen when
you write all your own tooling. Hopefully the more careful `.html`
file only filtering is the end of this.

Pull-request: https://github.com/evylang/evy/pull/433