labsite: Add rectangle challenge to ifs lab
Add rectangle challenge to ifs lab to provide an easy and open ended
exercise for students as the comparable "circle" challenge was one of
our most fun ones previously.

Add a small feature to lab notes that if an image has the
title "evy:edit" replace the title with "Click to edit" and load the
underlying Evy source into the editor on click. For example when
clicking on `samples/ifs/img/grass.svg` the sources of
`samples/ifs/img/grass.evy` are fetched an placed in the editor.