labsite: Change book-icon to back-icon for mobile
Change book-icon to back-icon for mobile: When on mobile only one of
these three screens is showed at a time: Notes, Code, Output. Clicking
the primary button lets you move from Notes -> Code <-> Output. To get
back to the Notes we had a book icon in the top right corner which
caused confusion. The back button is hopefully more logical and also
aligns with the whole screen swipe to to left.