📺 playsite: Remove slide back transition on navigation (#442)
When navigating to the next playground sample or lab challenge while on
the output view on mobile, we had this awkward: slide back to notes or
code first, than load new code. It seems more natural and intuitive to
just load the new contents - so for hashChange transition remove
animation by adding and eventually removing a CSS class that sets the
translate transition duration for the main container to 0 seconds.

Pull-request: https://github.com/evylang/evy/pull/442