play-site: Add Game of Life to interact samples
Add an implementation of Conway's Game of Life to the interact
samples. While this is titled "game", it is not really a game so does
not really belong in the games section. It features a few different
modes of interaction, so let's place it there.

Note: The Gen-command to update the changed snapshot also changed six
other images, likely due to slight differences in my local setup. I
reverted these six images and the tests still pass.

Gen-command: make snaps USE_DOCKER=1
Link: https://en.wikipedia.org/wiki/Conway%27s_Game_of_Life