🌌 play-site: Add Game of Life to interact samples (#446)
Add an implementation of Conway's Game of Life to the interact
samples. While this is titled "game", it is not really a game so does
not really belong in the games section. It features a few different
modes of interaction, so let's place it there.

Link: https://en.wikipedia.org/wiki/Conway%27s_Game_of_Life
Pull-request: https://github.com/evylang/evy/pull/446