make: Improve test-urls target
Improve test-urls target in Makefile to be more robust.
We are checking _all_ URLs in .md files for availability. This is an irregular task as we don't want to depend on other services / websites on CI.

We have update some markdown files to include YouTube embeds as HTML
tags. Inside these tags URLs are quotation mark enclosed, as is normal
in HTML. However, the regexp matching for URLs matched all characters
except for space. Add the quotation mark `"` to the list of characters
that do not form URLs, so that the regular expression now reads:

	https?://[^[:space:]"]+

This way, the URL will be matched correctly even in HTML tags.