lab: Add lab 5 - games
This lab is the transition from the foundation to the abstraction
course. It introduces the concept of events. We also show to program
simple games as this is something students have been asking for.

We've moved the unused bounce sample from the "if statement" lab.

The challenges are:

- 👋 Welcome
- 🎪 Events
- 🪙 Coin Toss
- 🎨 Paint
- 🎲 Guess
- 🟢 On animate
- 🏓 Bounce
- 👾 Game

It's all just new content, nothing has been added to the labs in terms
of features.