🐛 playsite: Fix scroll bug (#452)
Fix scroll bug where when an Evy program was started with ctrl+enter on
the playsite or labsite and there was an `on key` eventhandler handling
arrow key movements, the Evy code editor **also** started scrolling.

This behavior _only_ occurred when the Evy program was started with
ctrl+enter shortcut, not with the Run button.

Pull-request: https://github.com/evylang/evy/pull/452