labsite: Fix link to game solution
Fix link to game solution as the reference style link had duplicated
text and linked to previous partial solution where it should have
linked to the full solution in the end. Use more descriptive link
text.

This issue became apparent when revisiting the lab to extract the game
for the playground.