all: Rename "runtime" to "platform"
Rename uses of "runtime" where it refers to the CLI runtime and the
browser runtime to "platform". Runtime usually refers to the thing doing
the running, such as the evaluator or bytecode VM and has started to
cause a little confusion or cumbersome language now that the bytecode VM
is being developed. Initial talk of perhaps targeting micro-controllers
as "platforms" also suggests this should be called platform.

There are still a few uses of the word "runtime" left - these refer to
things like "runtime errors" where the term still makes sense, as it
does not refer to the Evy abstraction for runtime/platform.

Fix a lint error in `learn/pkg/learn/markdown.go` since golangci-lint
wants to complain about this now.

Commit-group: 1/2