svg: Prevent code with fused instructions
Add a floating point conversion when multiplying when scaling
coordinates so that the multiplication does not get fused with an
adjacent addition or subtraction. This causes two instructions to be
generated, a MUL and a ADD/SUB, instead of a single FMADD or FMSUB.
A fused instruction has different rounding so ends up giving slightly
different results to two instructions. This was causing different values
in the SVGs generated between AMD64 and ARM64.

Link: https://go.dev/ref/spec#Floating_point_operators