Add support for builtin functions
Begin small by adding the print and join, along with a rudimentary
concept of a Platform (runtime) copied from the evaluator package.
It would be ideal to extract the runtime into its own package, but
that felt out of scope for this change. Support was added in the
compiler to passthrough any types so that the argument to print
could be a parser.ANY_TYPE.